Unify rspack and webpack build-report plugins
